The anatomy of css syntax and CSS Selectors
The anatomy of CSS lecture added the original bacon fansite. In that lecture we linked the external css file and setup red text for H1 and the red background for the image.

For the CSS Selectors lecture, classes were introduced and we used classes to differentiate the background colors on the broccoli and bacon images.